+++ jdaviz/core/loaders/parsers/fits.py
@@ -13,7 +13,9 @@ class FITSParser(BaseParser):
 
     @property
     def is_valid(self):
-        if self.app.config not in ('deconfigged', 'specviz2d', 'lcviz', 'imviz', 'cubeviz'):
+        if self.app.config not in ('deconfigged', 'specviz2d',
+                                   'lcviz', 'imviz', 'cubeviz',
+                                   'rampviz'):
             # NOTE: temporary during deconfig process
             return False